people don't die anymore after the end of a game :)

+++ src/atomic-center/bomb.cpp
@@ -58,11 +58,11 @@ Bomb::explode()
 	if ( p )
 	{
 		Uint32 range = static_cast<Uint32>(p->getRange());
-		if ( map::Map::get(coords) == map::PLAYONBOMB )
+		if ( map::Map::get(coords) == map::PLAYONBOMB && ! Bomb::gameOver )
 			if ( p->kill(Player::playerAt(coords)) )
 				Bomb::gameOver = true;
 		bool up(true), down(true), right(true), left(true);
-		for ( Uint32 i = 1 ; i <= range ; ++i )
+		for ( Uint32 i(1) ; i <= range ; ++i )
 		{
 			if ( up )
 				up =    check(coords.x, coords.y - i);
